Ensure we clean up RuleSetFiles when we close the solution
The VisualStudioRuleSetManager is a singleton that acts as a cache
of RuleSetFiles and also manages the lifetime of file watching. We
never cleaned that up when we shut down VS, so we leaked a bunch of
file watchers, which would throw during finalization asserting they
weren't cleaned up and bring down the process with it.