Disable xUnit1013 in one file where it doesn't make sense
Right now this code is doing a pattern where it overrides the public
methods in derived types, and those have facts. That forces the base
type to also be public, but we can't put Facts there. This can be
cleaned up in a better way, but I want to keep this first round of
changes as mechanical as possible.