Enforce minimum MSBuild version
The Roslyn build depends on having a minimum of MSBuild 15.1 in order to
function. Anything less than that will fail with very crpytic errors,
usually around NuGet runtime identifiers.

This change causes our build to declaratively fail with a very specific
message if the wrong version of MSBuild is used. It should help point
customers in the right direction to fixing up the problem.

Note: This was made to be parameterized in Imports.targets because very
soon it will be changing to depend on 15.3.