Add null checks for TaskExtensions before deferred execution
A null value for certain arguments in TaskExtensions results in an exception.
We want this exception to reveal the caller who passed a null value. This
commit updates methods which defer the evaluation of the argument to explicitly
check for null early so the caller is revealed in the stack trace.