Don't pick a project arbitrarily when navigating to symbols
When we were navigating to a symbol with Go to Definition, we were
roundtripping the symbol to a symbol key, and then trying to
reconstitute the symbol by finding a project that had the reference
and then grabbing that symbol. There are two problems with this
approach:

1. We might get the symbol for the wrong language, which impacts the
   language we use to generate metadata as source. This fixes
   dotnet/roslyn#16818.

2. We might not even get the right symbol. Assembly identity in this
   case was using display name, which could be the wrong portable
   surface area or wrong version.

The fix is to instead carry along the project ID of the originating
symbol, and resolve against that.