Moved to xunit 2.2.0
This primarily moves us to the 2.2.0 version of xunit. In order to make
that transition though a number of other changes needed to happen as
well:

1. Needed to move our desktop target to 4.6.1. This should've been
required some time ago but we were essentially depending on bugs in the
SDK / NuGet.
1. Remove the DeployCoreClrTestRuntime project. This entire idea
depended on having the xunit.console.netcore package. This has since
been deleted and doesn't work with the 2.0 SDK.
1. Remove the deploy compiler tools project. Can just use dotnet exec
now and that's much more efficient.
1. Resolve a new ambiguity with KeyValuePair. A helper of ours now
exists in netcoreapp2.0.

I fully expect a number of Jenkins legs to fail with this change: build
correctness, coreclr and ubuntu. But I want to validate our other legs
and am still trackind down answers needed to fix the others.