Move command line tests back to net46
These projects reference the compiler EXE projects and hence copy them
to the output directory. This appears to not cause the correct
app.config file (ex csc.exe.config) to be copied along with it. It's
missing the necessary binding redirects that are generated from a direct
compile of the project. As a result when it' invoked from the unit tests
it fails.

It's possible this is a SDK bug and I will dig into that. For now though
I need to unblock the suites.