Fix a crash that  could occur in override completion
In order to insert code, override completion has to serialize/restore a
SymbolKey multiple times: once to actually create the completion item,
and additionally during the process of transforming the document.

Some symbolkeys cannot actually be restored. Specifically, a method
whose return type or parameters have use a type from an assembly not
referenced by the project the user is typing in cannot be restored. This
caused crashes/error dialogs when attempting override such members in a
way that is not predicable to the user.

To mitigate this we will:
* Not suggest members that can't be restored from their symbolkeys
* Stop the insertion process if we fail to restore a symbolkey

Fixes
https://devdiv.visualstudio.com/DefaultCollection/DevDiv/_workitems?id=419771&_a=edit