Update Developing a Language Feature.md
The "future" branch no longer exists. Now, there is a branch for each feature in development. Also, C# 6 is not the current version of the compiler anymore, as well as C# 7 is not the future version of the compiler anymore. I've adjusted this page in order to correspond to those changes accordingly.