Keep ProjectState._documentStates sorted rather than resorting the IDs
Since ImmutableDictionary and ImmutableSortedDictionary are both
implemented via trees; you can think of an ImmutableDictionary as just
being a sorted dictionary where the underlying key is the hash int.
Since our hash is just a hash of a GUID, we can compare by the GUID,
switch to a ImmutableSortedDictionary, and then avoid another sort
when we're computing checksums.