Ensure we always complete the call hierarchy search
If a symbol couldn't be resolved, we would bail the search and never
notify the UI that we stopped searching. We also wouldn't complete
the IAsyncToken either.

While I was here I also noticed that we were starting the async
operation asynchronously which could have resulted in flakiness in
integration tests. Our pattern is anything starting async work should
start the operation right away, so we can confidently wait until the
operation is complete.

# Conflicts:
#	src/EditorFeatures/Core/Implementation/CallHierarchy/Finders/AbstractCallFinder.cs