Ensure that we return the root IBlockOperation as the primary operation block for VB method bodies
Previously, we used to return two different operation blocks for the statements node within the body and the end block statement. This was incorrect, and was exposed when we recently started marking the implicit return statement in the sub as IsImplicit, and ended up with no explicit operations for an empty method body.