Remove CrossPlatform.sln (#22895)
The core work here is that we were pulling reference assemblies out of a folder with the name RoslynTools.ReferenceAssemblies. A while ago NuGet switched to unpacking all folders with lowercase paths. On Windows this continues to work fine as the filesystem APIs are case-insensitive, but this fails on Linux. By matching casing we can build both net461 and netcoreapp targets on any operating system. The inability to do this was the primary reason for CrossPlatform.sln, so there's now no reason for the solution and we can switch to using Compilers.sln, which is a more reasonable superset of everything in CrossPlatform.sln.

Fixes #20932