Ensure CodeModel adds elastic trivia when setting names
Fixes https://devdiv.visualstudio.com/DevDiv/_workitems/edit/512823

In VB Code Model, we use SyntaxFactory.ParseTypeName/ParseName when
setting the name of an attribute or namespace. Doing this fails to add
elastic trivia the way it is added in all other cases, so setting the
name of a namespace (for example) would remove the newline, creating
broken code. To fix this, we now manually add elastic trivia in these
cases.

C# was already functioning correctly, but it was using
".IdentifierName()" to parse dotted names, which isn't advised. This
change also updates the C# side accordingly.