Disable GenerateSatelliteAssemblies target
The standard `GenerateSatelliteAssemblies` target that ships with MSBuild uses al.exe (alink) to create satellite assemblies. However, there are a couple of things it doesn't support that we need in the dotnet/roslyn repo--see issues #23190 and #23191 for details. Instead, we want to use the `CoreGenerateSatelliteAssemblies` target from Microsoft.NET.Sdk. We've already opted in to that target, and now we need to opt out of the original `GenerateSatelliteAssemblies`. Until we can move to a copy of MSBuild that includes the change in https://github.com/Microsoft/msbuild/pull/2726 the best way to do that is to redefine the target as a no-op.