Decouple MiscellaneousFilesWorkspace from DocumentProvider
There was no deep reason to use DocumentProvider here, since the only
things it was being used for was to fetch the text buffer out of the
RDT (which MiscellaneousFilesWorkspace already did anyways) and to
apply text edits, which was just calling into one static function
anyways. Deleting this means DocumentProvider can be further refactored
and simplifed without this all getting in the way.