Sync compiler and vs toolset files
This changes the file set used to build the VS toolset to match that of
the standard compiler toolset. Simply re-use the file list in the
compiler nuspec file here.

This actuall found a bug in our current implementation. We were not
copying the config file for csi.exe in the VS toolset.