Make RootNamespace explicit in VB projects
Many of our VB projects today rely on having an empty default namespace
and explicitly specifying the full namespace in source. In the new SDK /
dir props world this can only be done by explicitly marking the
namespace as empty in the project file. Can't do the centralized control
that we do today.