Fix flaky compiler server test
MultipleSimultaneousCompiles calls RunCommandLineCompiler on multiple
threads (using Task.Run), which in turn calls CreateFiles on multiple
threads, which in turn goes through TempDirectory and TempRoot and
eventually boils down to calling List<T>.Add(), which is not
thread-safe, and so occasionally crashes.

Solution: hit the problem with a hammer by throwing a lock around CreateFiles.