Reorganize the central build logic
The combination of the minified MSBuild syntax and Directory.Build.props
/ targets file will change when in the build process our central build
files are executed. For instance Imports.targets will now be run much
later in the build after Sdk.targets. Where before, when it manually
invkoed Sdk.targets it was able to run both before and after it.

To keep our build functioning the same way we need to move around some
of the logic within our build files so that it executes in the same
relative order as it did before