Fix publish to include runtime identifier in dir
The publish directory needs to have the runtime identifier prepending to
the publish directory name. Previously out setting of `<OutputPath>`
happened early enough that the default logic kicked in. Now though it
happens too late. Copied the logic which used to set this for us and
filed a bug to undo all of this later.