Deploy Dir.build.props / targets in tests
Our MSBuild tests layout project / solution files on disk and then uses
MSBuild to evaluate them. During this evaluation MSBuild will search for
Directory.Build.props / targets files on the path just as it will for a
normal command line build.

On Jenkins our Directory.Build.props / targets is in the search path of
MSBuild and gets picked up during testing. This was causing our tests to
fail as we unconditionally override a number of properties in our core
targets files.

To insulate against this we now deploy and empty version into the
directory where we generate our solutions.