Update a unit test to reflect localized resources
One of the VB scripting tests changes the default culture to "de-DE" and
later prints out a string subject to localization. Now that we actually
have satellite assemblies with real translations where the runtime will
find them, we're getting the German version rather than the English
version.

The fix here is to simply update the baseline.