Add MEF metadata to IDeferredQuickInfoContentToFrameworkElementConverter
When I added this interface, I had the implementations expose their
applicable type not by metadata, but by simply calling a method on
the instance. This is problematic if the implementation is in a
different assembly as we'll load it too early. This is causing F#
to load when they shouldn't.

This is intended as a temporary shim that fixes the perf load for
the F# case (the only current exporter of these outside the Roslyn
repo itself) without forcing an interface change. This will get deleted
soon enough once we migrate to newer Quick Info APIs.