Correct date stamping logic for the new year
Ideally, we'd apply a version number to our binaries like 20180102
for the 2nd of January, 2018, but the Windows PE file format restricts
each number to a 16-bit unsigned integer. This means even trimming
to 80102 still isn't enough. The solution is to continue to decrement
that 'year' digit, and each time we do so add another 12 to the month.
This buys more time until we can rebaseline where we start.