Move to NetStandard2.0 RTM for our ref assemblies
This changes the reference assemblies we use for the CoreClr compiler
compilation to be Net Standard 2.0 RTM. Previously they were a beta
version.

This did mean a number of tests had to be moved to be desktop specific
again. This is primarily due to the presence of ValueTuple as a default
type in Net Standard 2.0. Tests that relied on ValueTuple not being
defined had to be moved to net461 which doesn't have that type defined
by default.