Report non-fatal errors if projects are removed that are referenced
We have seen some crash reports where we have a project being removed
that still has references pointing to it. This is mostly harmless but
sometimes can lead to downstream crashes. We've identified one bug
that we are fixing directly; this is to make further issues more
debuggable going forward.