Allow VB command-line arguments to add CorLibrary even if there are unresolved references
Previously, when calling `CommandLineArguements.ResolveMetadataReferences(...)` for VB would
not add the CorLibrary if any of the metadata references were unresolved. So, consumers would
need to remove any unresolved references prior to calling `CommandLineParser.Parse(...)` in
order to receive a valid metadata reference for mscorlib. This makes it very challenge to get a
correct set of references using the command-line compiler arguments produced by MSBuild.

This change updates `CommandLineArguments.ResolveMetadataReferences(...)` to compute the
CorLibrary even if there are unresolved metadata references.