Reconnect the persistence service to working folder location changes
This reimplements the entire mechanism for how we were passing around
location changes. The IPersistentStorageLocationService now also
has an event that it can raise to say the location has changed,
and it consumes this event. That part is actually the easy part of this
change.

More interesting is the persistence service is now more formally
defined to be a workspace-specific service that directly asks
for the IPersistentStorageLocationService upon creation. Previously
it was somewhat wishy-washy in that there was a singleton that
had a concept of the "global" primary solution, and if you passed
it a Solution from a different workspace that might work too, maybe.
Now the assumption is that a caller is using the service that matches
the solution they started with. This means the
IPersistentStorageLocationService is also a workspace service that
can be a non-singleton as well if different workspaces have
different concepts on how to get a working directory.

Since this required figuring out how some of the storage lifetime
worked, I decoupled the core implementations of IPersistentStorage
from having to also be concerned about lifetime. An implementation
is now standalone and Dispose means it should truely go away. The
service, since it's trying to manage a shared instance, does the
reference counting now and just uses ReferenceCountedDisposable
to manage that. This gives better protection too, in that if somebody
were to accidentally Dispose() twice we won't end up over-decrementing
the ref count and any number of bad things then happening.

Still to come is finishing up the remote workspace side of things.